Convert power parameters from double to int across SITypes API
This is a major API breaking change that affects all power-related functions
in the SITypes library to improve type safety and eliminate fractional
power ambiguities.

Major changes:
- SIDimensionality power functions now use int instead of double
- SIScalar power functions now use int instead of double
- SIUnit power functions now use int instead of double
- Added proper fractional power detection and rejection in parsers
- Updated all power operation tests to use integer powers
- Added debug output to help track unit parsing issues

API Changes:
- SIDimensionalityByRaisingToPower(dim, int power, error)
- SIDimensionalityByRaisingToPowerWithoutReducing(dim, int power, error)
- SIScalarCreateByRaisingToPower(scalar, int power, error)
- SIScalarRaiseToAPower(scalar, int power, error)
- SIUnitByRaisingToPower(unit, int power, multiplier, error)
- SIUnitByRaisingToPowerWithoutReducing(unit, int power, multiplier, error)

Breaking Change: Any code using fractional powers with these functions
will need to be updated to use the new SIScalarCreateByTakingNthRoot()
and SIUnitByTakingNthRoot() functions for fractional operations.

Fixes power operation ambiguities and improves API consistency.

+++ src/SIDimensionality.h
@@ -169,11 +169,11 @@ SIDimensionalityRef SIDimensionalityByDividing(SIDimensionalityRef theDim1, SIDi
 /** @brief Divides one dimensionality by another, does not reduce exponents. */
 SIDimensionalityRef SIDimensionalityByDividingWithoutReducing(SIDimensionalityRef theDim1, SIDimensionalityRef theDim2);
 
-/** @brief Raises a dimensionality to a real power, reduces exponents. */
-SIDimensionalityRef SIDimensionalityByRaisingToPower(SIDimensionalityRef theDim, double power, OCStringRef *error);
+/** @brief Create a new dimensionality by raising to a power and reducing exponents to lowest terms. */
+SIDimensionalityRef SIDimensionalityByRaisingToPower(SIDimensionalityRef theDim, int power, OCStringRef *error);
 
-/** @brief Raises a dimensionality to a real power, without reducing exponents. */
-SIDimensionalityRef SIDimensionalityByRaisingToPowerWithoutReducing(SIDimensionalityRef theDim, double power, OCStringRef *error);
+/** @brief Create a new dimensionality by raising to a power without reducing exponents. */
+SIDimensionalityRef SIDimensionalityByRaisingToPowerWithoutReducing(SIDimensionalityRef theDim, int power, OCStringRef *error);
 
 /** @brief Returns all quantity names matching this dimensionality exactly. */
 OCArrayRef SIDimensionalityCreateArrayOfQuantities(SIDimensionalityRef theDim);
